Add optional chaining to Instance_UUID and Schema_UUID
Ensured that a call to find on the metrics array in mqttclient.ts will not raise an error if there are no matching records. Previously, if no metric with name "Instance_UUID" or "Schema_UUID" was found, a TypeError would be thrown while attempting to access the 'value' property. The application to use optional chaining (?.) so it will return undefined rather than throwing an error.